moved some functions
