+++ rust_everscale_tezos/tezos_everscale/dependencies/lib/src/everscale/htlc.rs
@@ -1,9 +1,9 @@
 use ton_client::ClientContext;
 use serde_json::Value;
 use std::sync::Arc;
-use crate::get::*;
 use super::contract::call_contract_with_client;
 use super::config::Config;
+use super::get::*;
 
 pub async fn create_lock_with_tokens(
     ton: Arc<ClientContext>,
