+++ rust_everscale_tezos/tezos_everscale/dependencies/lib/src/tezos/operation.rs
@@ -1,12 +1,11 @@
-use crate::common::operation_command::exit_with_error_no_wallet_type_selected;
-use crate::common::operation_command::LocalWalletState;
-use crate::tezos::get::*;
-use crate::get::*;
-use crate::*;
+use super::get::*;
 use rpc_api::api::InjectOperationsResult;
 use rpc_api::api::RunOperationError;
 use signer::OperationSignatureInfo;
 use hex::FromHex;
+use crate::common::operation_command::exit_with_error_no_wallet_type_selected;
+use crate::common::operation_command::LocalWalletState;
+use crate::*;
 
 pub struct OperationResult {
     pub consumed_gas: String,
