+++ rust_everscale_tezos/tezos_everscale/src/bin/60_relay_xtz_to_wxtz.rs
@@ -1,4 +1,9 @@
-use lib::everscale::{config::{Config, create_client_verbose}, get::{ever_get_transactions, get_payload}, multisig::{submit_transaction, confirm_transaction}, htlc::{get_timestamp, create_lock_with_tokens}};
+use lib::everscale::htlc::create_lock_with_tokens;
+use lib::everscale::htlc::get_timestamp;
+use lib::everscale::multisig::confirm_transaction;
+use lib::everscale::multisig::submit_transaction;
+use lib::everscale::config::*;
+use lib::everscale::get::*;
 use serde_json::Value;
 use lib::functions::*;
 use ureq::Agent;
